backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-docx>=0.8.11
httpx>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
openai>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0